[varLib cff2] Sparse source support. Bug fix in models doc test
Fixed doc test to match new behavior after bug fix in _computeMasterSupports() at line 296.

I guess pytest does not exercise the doc tests: I  didn't see this when running pytest on my local system.

Before bug fix, support for master 7 was " {'wdth': (0, 0.3, 0.3), 'wght': (0, 120, 180)},", and it now is "{'wdth': (0.2, 0.3, 0.3), 'wght': (0, 120, 180)}". The low value for the support (low, peak, high) for master 7 should be next lowest peak on the same axis, which is 0.2, not 0.